% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locked_list.r
\name{lockedlist}
\alias{lockedlist}
\title{lockedlist class}
\description{
persists a list over filesystem and provides thread safe
access via a file lock
}
\examples{
path <- rutils::tempdir()
on.exit(unlink(path, recursive = TRUE))
ll <- lockedlist$new(path)
ll$set(list(A = 1))
x <- ll$get("A")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LockedList-new}{\code{lockedlist$new()}}
\item \href{#method-LockedList-get}{\code{lockedlist$get()}}
\item \href{#method-LockedList-set}{\code{lockedlist$set()}}
\item \href{#method-LockedList-print}{\code{lockedlist$print()}}
\item \href{#method-LockedList-names}{\code{lockedlist$names()}}
\item \href{#method-LockedList-clone}{\code{lockedlist$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LockedList-new"></a>}}
\if{latex}{\out{\hypertarget{method-LockedList-new}{}}}
\subsection{Method \code{new()}}{
Creates a lockedlist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lockedlist$new(path = rutils::tempdir("LockedList"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path where data will be serialized
(defaults to \code{rutils::tempdir("LockedList")})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LockedList-get"></a>}}
\if{latex}{\out{\hypertarget{method-LockedList-get}{}}}
\subsection{Method \code{get()}}{
retrievs data based on names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lockedlist$get(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string identifing the object in the \code{data} list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LockedList-set"></a>}}
\if{latex}{\out{\hypertarget{method-LockedList-set}{}}}
\subsection{Method \code{set()}}{
Set a list in this lockedlist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lockedlist$set(obj_as_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj_as_list}}{a names list object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LockedList-print"></a>}}
\if{latex}{\out{\hypertarget{method-LockedList-print}{}}}
\subsection{Method \code{print()}}{
print a string repr of this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lockedlist$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{generic params (unused)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LockedList-names"></a>}}
\if{latex}{\out{\hypertarget{method-LockedList-names}{}}}
\subsection{Method \code{names()}}{
yields the names of the \code{data} list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lockedlist$names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LockedList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LockedList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lockedlist$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
